requirements.txt
streamlit
opencv-python-headless
torch
torchvision
pillow
numpy